+++ models/confirmation.go
@@ -23,7 +23,7 @@ type (
 		UserId       string       `json:"-" bson:"userId"`
 		Team         *Team        `json:"target" bson:",inline"`
 		Role         string       `json:"role" bson:"role"`
-		Status       Status       `json:"-" bson:"status"`
+		Status       Status       `json:"status" bson:"status"`
 		Modified     time.Time    `json:"-" bson:"modified"`
 		ShortKey     string       `json:"shortKey" bson:"shortKey"`
 	}
@@ -88,11 +88,12 @@ const (
 
 var (
 	Timeouts TypeDurations = TypeDurations{
-		TypeCareteamInvite:       7 * 24 * time.Hour,
-		TypePasswordReset:        7 * 24 * time.Hour,
-		TypeSignUp:               31 * 24 * time.Hour,
-		TypePatientPasswordReset: 1 * time.Hour,
-		TypeMedicalTeamInvite:    7 * 24 * time.Hour,
+		TypeCareteamInvite:           7 * 24 * time.Hour,
+		TypePasswordReset:            7 * 24 * time.Hour,
+		TypeSignUp:                   31 * 24 * time.Hour,
+		TypePatientPasswordReset:     1 * time.Hour,
+		TypeMedicalTeamInvite:        7 * 24 * time.Hour,
+		TypeMedicalTeamPatientInvite: 7 * 24 * time.Hour,
 	}
 )